feat: add remote prop
