--- docs/demos/select/persist/desc.en-US.md
@@ -1,7 +1,0 @@
-### Persist Options
-
-Add the `filter` prop to enable option filter.
-
-By default, the built-in filter method is used. If you want to customize the filter method, you can pass a function.
-
-Add the `ignore-case` prop to ignore case when using built-in filter methods.
