--- docs/demos/select/persist/desc.zh-CN.md
@@ -1,7 +1,0 @@
-### 持久化选项
-
-添加 `filter` 属性可以开启选项过滤。
-
-默认使用内置的过滤方法，如果你想要自定义过滤的方法可以传入一个函数。
-
-添加 `ignore-case` 属性可以在使用内置过滤方法时忽略大小写。
