+++ docs/demos/select/remote/desc.en-US.md
@@ -1,7 +1,5 @@
-### Filter Options
+### Remote Mode
 
-Add the `filter` prop to enable option filter.
+In normal mode, values not included in the options will be ignored.
 
-By default, the built-in filter method is used. If you want to customize the filter method, you can pass a function.
-
-Add the `ignore-case` prop to ignore case when using built-in filter methods.
+Add the `remote` props to enable remote mode, in which values will be cached and filtered to only dispatch events.
