+++ docs/demos/select/remote/desc.zh-CN.md
@@ -1,7 +1,5 @@
-### 过滤选项
+### 远程模式
 
-添加 `filter` 属性可以开启选项过滤。
+普通模式下，选项中未包含的值将会被忽略。
 
-默认使用内置的过滤方法，如果你想要自定义过滤的方法可以传入一个函数。
-
-添加 `ignore-case` 属性可以在使用内置过滤方法时忽略大小写。
+添加 `remote` 属性可以开启远程模式，在该模式下值将被缓存，过滤仅派发事件。
